This PR make n_threads and n_threads_batch configurable in batched & batched-bench example.

- Add n_threads/n_threads_batch option to batched
- Add n_threads/n_threads_batch option to batched-bench
- Update those README.md

The motivation for this is that it can be used to check how many threads the performance scales in batch inference on many cores systems; it may be useful to determine how many cores to allocate for k8s pod or VM instance.