@msy-kato 2024-05-28
This is logs

<details>
<summary>batched example</summary>

```bash
### batched
$ ./batched models/llama-2-7b-chat.Q8_0.gguf 'Hello my name is' 4 128 0 16 16
(...snip)
llm_load_print_meta: vocab type       = SPM                                                                                                                                                                                                                                                          
llm_load_print_meta: n_vocab          = 32000                                                                                                                                                                                                                                                        
llm_load_print_meta: n_merges         = 0                                                                                                                                                                                                                                                            
llm_load_print_meta: n_ctx_train      = 2048                                                                                                                                                                                                                                                         
llm_load_print_meta: n_embd           = 4096                                                                                                                                                                                                                                                         
llm_load_print_meta: n_head           = 32                                                                                                                                                                                                                                                           
llm_load_print_meta: n_head_kv        = 32                                                                                                                                                                                                                                                           
llm_load_print_meta: n_layer          = 32                                                                                                                                                                                                                                                           
llm_load_print_meta: n_rot            = 128                                                                                                                                                                                                                                                          
llm_load_print_meta: n_embd_head_k    = 128                                                                                                                                                                                                                                                          
llm_load_print_meta: n_embd_head_v    = 128                                                                                                                                                                                                                                                          
llm_load_print_meta: n_gqa            = 1                                                                                                                                                                                                                                                            
llm_load_print_meta: n_embd_k_gqa     = 4096                                                                                                                                                                                                                                                         
llm_load_print_meta: n_embd_v_gqa     = 4096                                                                                                                                                                                                                                                         
llm_load_print_meta: f_norm_eps       = 0.0e+00                                                                                                                                                                                                                                                      
llm_load_print_meta: f_norm_rms_eps   = 1.0e-06                                                                                                                                                                                                                                                      
llm_load_print_meta: f_clamp_kqv      = 0.0e+00                                                                                                                                                                                                                                                      
llm_load_print_meta: f_max_alibi_bias = 0.0e+00                                                                                                                                                                                                                                                      
llm_load_print_meta: f_logit_scale    = 0.0e+00                                                                                                                                                                                                                                                      
llm_load_print_meta: n_ff             = 11008                                                                                                                                                                                                                                                        
llm_load_print_meta: n_expert         = 0                                                                                                                                                                                                                                                            
llm_load_print_meta: n_expert_used    = 0                                                                                                                                                                                                                                                            
llm_load_print_meta: causal attn      = 1                                                                                                                                                                                                                                                            
llm_load_print_meta: pooling type     = 0                                                                                                                                                                                                                                                            
llm_load_print_meta: rope type        = 0                                                                                                                                                                                                                                                            
llm_load_print_meta: rope scaling     = linear                                                                                                                                                                                                                                                       
llm_load_print_meta: freq_base_train  = 10000.0                                                                                                                                                                                                                                                      
llm_load_print_meta: freq_scale_train = 1                                                                                                                                                                                                                                                            
llm_load_print_meta: n_yarn_orig_ctx  = 2048                                                                                                                                                                                                                                                         
llm_load_print_meta: rope_finetuned   = unknown                                                                                                                                                                                                                                                      
llm_load_print_meta: ssm_d_conv       = 0
llm_load_print_meta: ssm_d_inner      = 0
llm_load_print_meta: ssm_d_state      = 0
llm_load_print_meta: ssm_dt_rank      = 0
llm_load_print_meta: model type       = 7B
llm_load_print_meta: model ftype      = Q8_0
llm_load_print_meta: model params     = 6.74 B
llm_load_print_meta: model size       = 6.67 GiB (8.50 BPW)
llm_load_print_meta: general.name     = llama
llm_load_print_meta: BOS token        = 1 '<s>'
llm_load_print_meta: EOS token        = 2 '</s>'
llm_load_print_meta: UNK token        = 0 '<unk>'
llm_load_print_meta: LF token         = 13 '<0x0A>'
llm_load_tensors: ggml ctx size =    0.15 MiB
llm_load_tensors:        CPU buffer size =  6828.64 MiB
...................................................................................................
llama_new_context_with_model: n_ctx      = 512
llama_new_context_with_model: n_batch    = 128
llama_new_context_with_model: n_ubatch   = 128
llama_new_context_with_model: flash_attn = 0
llama_new_context_with_model: freq_base  = 10000.0
llama_new_context_with_model: freq_scale = 1
llama_kv_cache_init:        CPU KV buffer size =   256.00 MiB
llama_new_context_with_model: KV self size  =  256.00 MiB, K (f16):  128.00 MiB, V (f16):  128.00 MiB
llama_new_context_with_model:        CPU  output buffer size =     0.49 MiB
llama_new_context_with_model:        CPU compute buffer size =    17.63 MiB
llama_new_context_with_model: graph nodes  = 1030
llama_new_context_with_model: graph splits = 1

main: n_len = 128, n_ctx = 512, n_batch = 128, n_parallel = 4, n_kv_req = 497, n_threads = 16, n_threads_batch = 16

<s> Hello my name is

main: generating 4 sequences ...

main: stream 0 finished at n_cur = 128
main: stream 1 finished at n_cur = 128
main: stream 2 finished at n_cur = 128
main: stream 3 finished at n_cur = 128

sequence 0:

Hello my name is [Your Name], and I am a [Your Profession] with [Number of Years] of experience in the [Your Industry/Field] field. I am excited to be here today to share my insights on [Topic/Theme].

In my [Number of Years] of experience, I have had the opportunity to work with a diverse range of clients, from small startups to large corporations, and have seen firsthand the impact that [Topic/Theme] can have on their businesses. Through my work, I have developed a deep understanding of the challenges and opportun

sequence 1:

Hello my name is Drew and I am a 31 year old man from the United States. I have been a fan of anime for as long as I can remember. I have been watching anime since I was a kid and have always been fascinated by the unique art style, storytelling, and characters found in this medium.
As I grew older, my interest in anime only grew stronger. I began to explore different genres and styles, from action-packed series like "Dragon Ball Z" and "Naruto" to more emotional and thought-provoking shows like

sequence 2:

Hello my name is Tiffany and I am a 30 year old female. I have been experiencing some symptoms that I am concerned about and I am not sure what to do. I have been feeling really tired and sluggish all the time, even after getting a full night's sleep. I also have been experiencing some abdominal pain and bloating, and I have been having some trouble with my periods. I am also noticing some changes in my skin, such as acne and dark spots. I am not sure if these symptoms are related, but I am worried that

sequence 3:

Hello my name is John and I am a 26 year old man from the United States. I have been experiencing some strange symptoms over the past few months that I am concerned about. I have been having trouble sleeping at night and have been feeling very tired during the day. I have also been experiencing some muscle weakness and stiffness, especially in my legs. I have been experiencing some pain in my lower back and hips as well. I have been to see my doctor and they have run some tests, but they have not been able to find anything wrong with me. I am starting

main: decoded 492 tokens in 37.96 s, speed: 12.96 t/s

llama_print_timings:        load time =     784.66 ms
llama_print_timings:      sample time =      17.97 ms /   496 runs   (    0.04 ms per token, 27600.02 tokens per second)
llama_print_timings: prompt eval time =   38272.15 ms /   497 tokens (   77.01 ms per token,    12.99 tokens per second)
llama_print_timings:        eval time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
```

</details>

<details>
<summary>batched-bench example</summary>

```bash
### batched-bench
$ ./batched-bench models/llama-2-7b-chat.Q8_0.gguf 16384 2048 512 0 0 999 32 32 128,256,512 128,256 1,2,4,8,16,32
(...snip)
llm_load_vocab: special tokens definition check successful ( 259/32000 ).
llm_load_print_meta: format           = GGUF V3 (latest)
llm_load_print_meta: arch             = llama
llm_load_print_meta: vocab type       = SPM
llm_load_print_meta: n_vocab          = 32000
llm_load_print_meta: n_merges         = 0
llm_load_print_meta: n_ctx_train      = 2048
llm_load_print_meta: n_embd           = 4096
llm_load_print_meta: n_head           = 32
llm_load_print_meta: n_head_kv        = 32
llm_load_print_meta: n_layer          = 32
llm_load_print_meta: n_rot            = 128
llm_load_print_meta: n_embd_head_k    = 128
llm_load_print_meta: n_embd_head_v    = 128
llm_load_print_meta: n_gqa            = 1
llm_load_print_meta: n_embd_k_gqa     = 4096
llm_load_print_meta: n_embd_v_gqa     = 4096
llm_load_print_meta: f_norm_eps       = 0.0e+00
llm_load_print_meta: f_norm_rms_eps   = 1.0e-06
llm_load_print_meta: f_clamp_kqv      = 0.0e+00
llm_load_print_meta: f_max_alibi_bias = 0.0e+00
llm_load_print_meta: f_logit_scale    = 0.0e+00
llm_load_print_meta: n_ff             = 11008
llm_load_print_meta: n_expert         = 0
llm_load_print_meta: n_expert_used    = 0
llm_load_print_meta: causal attn      = 1
llm_load_print_meta: pooling type     = 0
llm_load_print_meta: rope type        = 0
llm_load_print_meta: rope scaling     = linear
llm_load_print_meta: freq_base_train  = 10000.0
llm_load_print_meta: freq_scale_train = 1
llm_load_print_meta: n_yarn_orig_ctx  = 2048
llm_load_print_meta: rope_finetuned   = unknown
llm_load_print_meta: ssm_d_conv       = 0
llm_load_print_meta: ssm_d_inner      = 0
llm_load_print_meta: ssm_d_state      = 0
llm_load_print_meta: ssm_dt_rank      = 0
llm_load_print_meta: model type       = 7B
llm_load_print_meta: model ftype      = Q8_0
llm_load_print_meta: model params     = 6.74 B
llm_load_print_meta: model size       = 6.67 GiB (8.50 BPW)
llm_load_print_meta: general.name     = llama
llm_load_print_meta: BOS token        = 1 '<s>'
llm_load_print_meta: EOS token        = 2 '</s>'
llm_load_print_meta: UNK token        = 0 '<unk>'
llm_load_print_meta: LF token         = 13 '<0x0A>'
llm_load_tensors: ggml ctx size =    0.15 MiB
llm_load_tensors:        CPU buffer size =  6828.64 MiB
...................................................................................................
llama_new_context_with_model: n_ctx      = 16384
llama_new_context_with_model: n_batch    = 2048
llama_new_context_with_model: n_ubatch   = 512
llama_new_context_with_model: flash_attn = 0
llama_new_context_with_model: freq_base  = 10000.0
llama_new_context_with_model: freq_scale = 1
llama_kv_cache_init:        CPU KV buffer size =  8192.00 MiB
llama_new_context_with_model: KV self size  = 8192.00 MiB, K (f16): 4096.00 MiB, V (f16): 4096.00 MiB
llama_new_context_with_model:        CPU  output buffer size =     3.91 MiB
llama_new_context_with_model:        CPU compute buffer size =  1088.01 MiB
llama_new_context_with_model: graph nodes  = 1030
llama_new_context_with_model: graph splits = 1

main: n_kv_max = 16384, n_batch = 2048, n_ubatch = 512, flash_attn = 0, is_pp_shared = 0, n_gpu_layers = 999, n_threads = 32, n_threads_batch = 32

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |    128 |    1 |    256 |    4.636 |    27.61 |    8.188 |    15.63 |   12.824 |    19.96 |
|   128 |    128 |    2 |    512 |    9.188 |    27.86 |   12.596 |    20.32 |   21.784 |    23.50 |
|   128 |    128 |    4 |   1024 |   18.553 |    27.60 |   22.736 |    22.52 |   41.289 |    24.80 |
|   128 |    128 |    8 |   2048 |   37.678 |    27.18 |   44.471 |    23.03 |   82.149 |    24.93 |
|   128 |    128 |   16 |   4096 |   78.098 |    26.22 |   95.607 |    21.42 |  173.705 |    23.58 |
```
</details>